requirements.txt
python-pptx
streamlit
python-docx
google-generativeai
sentence-transformers
numpy
pandas
scikit-learn